'Initial_Commit'

+++ source/i2c.h
@@ -10,7 +10,7 @@ typedef enum
 }I2CStatus_t;
 
 void i2cInit(void);
-I2CStatus_t i2cSetTLow(uint32_t t_low);
-I2CStatus_t I2CReadTemperature(volatile uint32_t * raw_temp);
+I2CStatus_t i2cSetTLow(int16_t t_low);
+I2CStatus_t I2CReadTemperature(volatile int16_t * raw_temp);
 
 #endif /* _I2C_H */
